+++ Modules/Intent.Modules.Integration.HttpClients/Templates/HttpClient/HttpClientTemplateRegistration.cs
@@ -40,7 +40,8 @@ namespace Intent.Modules.Integration.HttpClients.Templates.HttpClient
             const string serviceProxiesDesignerId = "2799aa83-e256-46fe-9589-b96f7d6b09f7";
             return _metadataManager.GetServiceProxyModels(
                 application.Id,
-                applicationId => _metadataManager.GetDesigner(applicationId, serviceProxiesDesignerId), // for backward compatibility
+                application,
+                appId => _metadataManager.GetDesigner(appId, serviceProxiesDesignerId), // for backward compatibility
                 _metadataManager.Services);
         }
     }