+++ Modules/Intent.Modules.Integration.HttpClients/release-notes.md
@@ -1,6 +1,7 @@
 ### Version 6.1.0
 
 - Improvement: Updated Dependencies and versions.
+- Fixed: Services secured by default will now correctly propagate to HttpClients.
 
 ### Version 6.0.14